Diet/requirements.txt
anthropic==0.29.0
boto3==1.34.131
botocore==1.34.131
gradio==4.36.1
gradio_client==1.0.1
numpy==2.0.0
openai==1.35.3
pillow==10.3.0
requests==2.32.3
watchdog==4.0.1
watchfiles==0.22.0

(must also setup AWS CLI)